requirements.txt
alembic==1.6.5
anyio==3.3.0
appdirs==1.4.4
APScheduler==3.7.0
autopep8==1.5.7
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
distlib==0.3.2
docker==5.0.0
filelock==3.0.12
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.25
pipenv==2021.5.29
protobuf==3.17.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
rfc3986==1.5.0
scrapinghub==2.3.1
six==1.16.0
sniffio==1.2.0
spotipy==2.18.0
SQLAlchemy==1.4.20
toml==0.10.2
tqdm==4.55.1
tzlocal==2.1
urllib3==1.26.6
uWSGI==2.0.19.1
virtualenv==20.4.7
virtualenv-clone==0.5.6
websocket-client==1.1.1
Werkzeug==2.0.1
youtube-search-python==1.4.6
